requirements.txt
Django>=3.0.2,<3.1.0
djangorestframework>=3.11.0,<3.12.0
djoser>=2.0.3,<2.1.0
Pillow>=7.0.0,<7.1.0
drf_yasg>=1.17.1,<1.18.0
